Changed preload to load
Current bug: When interacting with an item box to spawn an item, if the player picks this new item up immediately, it picks up the object they last picked up instead. If they wait a little bit before picking up the new item, it works as intended. Not sure why!